Restrict attribute matching to used attributes
This reverts an earlier change which made attribute matching
consider all attributes in the schema to avoid creating a
new set of attributes on each match. This used slightly less
memory, but was more CPU intensive, as many matches only use
a subset of the attributes in the schema. This chagne goes
back to the original behavior and uses an immutable set,
which is more memory friendly than the original hashset
implementation, getting the best of both worlds.